Refactoring: Move filters to search package, rename Searchable to SearchableTableQuery
