+++ src/test/scala/za/co/absa/hyperdrive/trigger/models/filters/IntRangeFilterTest.scala
@@ -16,6 +16,7 @@
 package za.co.absa.hyperdrive.trigger.models.filters
 
 import org.scalatest.{FlatSpec, _}
+import za.co.absa.hyperdrive.trigger.models.search.IntRangeFilterAttributes
 
 class IntRangeFilterTest extends FlatSpec with Matchers with BeforeAndAfterAll with BeforeAndAfterEach with FilterTestBase {
   import h2Profile.api._
