+++ src/test/scala/za/co/absa/hyperdrive/trigger/models/filters/StringEqualsFilterTest.scala
@@ -16,6 +16,7 @@
 package za.co.absa.hyperdrive.trigger.models.filters
 
 import org.scalatest.{FlatSpec, _}
+import za.co.absa.hyperdrive.trigger.models.search.StringEqualsFilterAttributes
 
 class StringEqualsFilterTest extends FlatSpec with Matchers with BeforeAndAfterAll with BeforeAndAfterEach with FilterTestBase {
   import h2Profile.api._
